swal added to delete and edit

+++ client/src/components/Comments.vue
@@ -8,6 +8,7 @@
 </template>
 
 <script>
+import Swal from "sweetalert2";
 export default {
   name: "Comment",
   
@@ -15,7 +16,21 @@ export default {
 
   methods: {
     removeComment(comment) {
-      this.$store.dispatch("deleteComment", comment);
+
+    Swal.fire({
+        title: "Are you sure you want to delete this comment?",
+        text: "You will not be able to undo this delete.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085D6",
+        cancelButtonColor: "#D33",
+        confirmButtonText: "Yes, Delete Comment"
+      }).then(result => {
+        if (result.value) {
+          this.$store.dispatch("deleteComment", comment);
+          Swal.fire("Deleted!", "Your comment has been deleted.", "Success");
+        }
+      });
     }
   }
 };
